RVV: move rvv_helper/rvv_traits to gdalrvv.hpp
